requirements.txt
# Core web framework
quart>=0.18.0

# Poe API client
fastapi-poe>=0.0.20

# Configuration management
pyyaml>=6.0

# Async file operations
aiofiles>=23.1.0

# Rate limiting
quart-rate-limiter>=0.7.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Security
python-dotenv>=1.0.0
bcrypt>=4.0.1

# Monitoring
prometheus-client>=0.17.0